// prisma/generated/schemas/objects/NestedFloatNullableFilter.schema.ts
import type { Prisma } from '@prisma/client'
import { z } from 'zod'

const Schema: z.ZodType<Prisma.NestedFloatNullableFilter> = z
  .object({
    equals: z.number().optional().nullable(),
    in: z.number().array().optional().nullable(),
    notIn: z.number().array().optional().nullable(),
    lt: z.number().optional(),
    lte: z.number().optional(),
    gt: z.number().optional(),
    gte: z.number().optional(),
    not: z
      .union([z.number(), z.lazy(() => NestedFloatNullableFilterObjectSchema)])
      .optional()
      .nullable(),
  })
  .strict()

export const NestedFloatNullableFilterObjectSchema = Schema